+++ integration_tests/codegen_fail/fail/scalar_value/derive_on_struct.rs
@@ -1,4 +1,4 @@
-#[derive(juniper::GraphQLScalarValue)]
+#[derive(juniper::ScalarValue)]
 struct ScalarValue;
 
 fn main() {}