+++ integration_tests/codegen_fail/fail/scalar_value/multiple_named_fields.rs
@@ -1,4 +1,4 @@
-#[derive(juniper::GraphQLScalarValue)]
+#[derive(juniper::ScalarValue)]
 enum ScalarValue {
     Variant { first: i32, second: u64 },
 }