+++ integration_tests/codegen_fail/fail/scalar_value/multiple_unnamed_fields.rs
@@ -1,4 +1,4 @@
-#[derive(juniper::GraphQLScalarValue)]
+#[derive(juniper::ScalarValue)]
 enum ScalarValue {
     Variant(u32, i64),
 }